2021.7.9 0:04

+++ frontend/src/components/SubmitForm.js
@@ -110,21 +110,36 @@ class SubmitForm extends React.Component {
         const callback = (data) => {
             console.log("call  " + data);
         }
-        orderService.addOrder(json, callback);
-        let json2 = new Object();
-        let delArray = [];
-        for(let i = 0; i < global.checked.length; i++){
-            let book_id = global.buy[i].book_id;
-            console.log("global.buy[i]  " + book_id);
-            console.log(global.buy);
-            delArray.push(book_id);
+        if(global.orderFrom == "cart"){
+            orderService.addOrder(json, callback);
+            let json2 = new Object();
+            let delArray = [];
+            for(let i = 0; i < global.checked.length; i++){
+                let book_id = global.buy[i].book_id;
+                console.log("global.buy[i]  " + book_id);
+                console.log(global.buy);
+                delArray.push(book_id);
+            }
+            json2.id = delArray;
+            json2.user_id = user_id;
+            const callback2 = (data) => {
+                console.log("call  " + data);
+            }
+            cartService.deleteItem(json2, callback2);
         }
-        json2.id = delArray;
-        json2.user_id = user_id;
-        const callback2 = (data) => {
-            console.log("call  " + data);
+        else{
+            let user_id = this.state.user;
+            let json = new Object();
+            json.user_id = user_id;
+            json.totalmoney = global.buy[0].price;
+            json.book_id = global.buy[0].id;
+            console.log(global.buy[0])
+            json.num = 1;
+            const callback = (data) => {
+                console.log("call  " + data);
+            }
+            orderService.addOrderOne(json, callback);
         }
-        cartService.deleteItem(json2, callback2);
     }
 
     // handleClick = e =>{
@@ -138,7 +153,6 @@ class SubmitForm extends React.Component {
         console.log("buy", global.buy);
         add();
         console.log(dataSource);
-
         const { getFieldDecorator } = this.props.form;
         return (
             <div>
